requirements.txt
pyperclip
pillow
python-magic
